+++ trace-processor/common/src/main/java/eu/csgroup/coprs/monitoring/common/ingestor/EntityIngestor.java
@@ -141,7 +141,15 @@ public class EntityIngestor implements EntityFinder {
         return (EntityRepository<T, E>) specializedRepository;
     }
 
-
+    /**
+     * Store in database list of entities (can be a list containing different entity type). Entity storage is done
+     * in a certain order:
+     * <br>
+     * First Entities that don't have foreign keys towards other entities, then the ones that do.
+     *
+     * @param entities processed by ProcessorOrchestrator
+     * @return The entities saved in database
+     */
     public List<DefaultEntity> saveAll(List<DefaultEntity> entities) {
         if (entities == null || entities.isEmpty()) {
             return List.of();
@@ -182,7 +190,7 @@ public class EntityIngestor implements EntityFinder {
         var repo = selectRepository(entry.getKey());
         timer.startUnitaryTimer(entry.getKey());
         var defaultEntities = repo.saveAll(entry.getValue());
-        timer.endtUnitaryTimer(entry.getKey());
+        timer.endUnitaryTimer(entry.getKey());
         return defaultEntities.stream();
     }